+++ browser/components/newtab/karma.mc.config.js
@@ -136,17 +136,39 @@ module.exports = function (config) {
               functions: 100,
               branches: 75,
             },
-            "lib/TopSitesFeed.jsm": {
-              statements: 70,
-              lines: 75,
-              functions: 80,
-              branches: 60,
-            },
             "lib/Screenshots.jsm": {
               statements: 94,
               lines: 94,
               functions: 75,
               branches: 84,
+            },
+            /**
+             * Store.jsm is tested via an xpcshell test
+             */
+            "lib/Store.jsm": {
+              statements: 8,
+              lines: 8,
+              functions: 0,
+              branches: 0,
+            },
+            /**
+             * TopSitesFeed.jsm is tested via an xpcshell test
+             */
+            "lib/TopSitesFeed.jsm": {
+              statements: 9,
+              lines: 9,
+              functions: 5,
+              branches: 0,
+            },
+            /**
+             * TopStoresFeed.jsm is not tested in automation and is slated
+             * for eventual removal.
+             */
+            "lib/TopStoriesFeed.jsm": {
+              statements: 0,
+              lines: 0,
+              functions: 0,
+              branches: 0,
             },
             "lib/*.jsm": {
               statements: 100,
@@ -184,45 +206,11 @@ module.exports = function (config) {
               functions: 0,
               branches: 0,
             },
-            "content-src/lib/aboutwelcome-utils.js": {
-              statements: 50,
-              lines: 50,
-              functions: 50,
-              branches: 0,
-            },
             "content-src/lib/link-menu-options.js": {
               statements: 96,
               lines: 96,
               functions: 96,
               branches: 70,
-            },
-            "content-src/aboutwelcome/components/LanguageSwitcher.jsx": {
-              // This file is covered by the mochitest: browser_aboutwelcome_multistage_languageSwitcher.js
-              statements: 0,
-              lines: 0,
-              functions: 0,
-              branches: 0,
-            },
-            "content-src/aboutwelcome/components/EmbeddedMigrationWizard.jsx": {
-              // This file is covered by the mochitest: browser_aboutwelcome_multistage_mr.js
-              // Can't be unit tested because it relies on the migration-wizard custom element
-              statements: 0,
-              lines: 0,
-              functions: 0,
-              branches: 0,
-            },
-            "content-src/aboutwelcome/components/AddonsPicker.jsx": {
-              // This file is covered by the mochitest: browser_aboutwelcome_multistage_addonspicker.js
-              statements: 0,
-              lines: 0,
-              functions: 0,
-              branches: 0,
-            },
-            "content-src/aboutwelcome/**/*.jsx": {
-              statements: 62,
-              lines: 60,
-              functions: 50,
-              branches: 50,
             },
             "content-src/components/**/*.jsx": {
               statements: 51.1,